BreastCancerDetector_simulation/requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
click==8.1.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.1
fonttools==4.32.0
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pipp==0.0.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
setuptools-scm==6.4.2
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
